+++ packages/std/src/lib.rs
@@ -23,8 +23,7 @@ mod timestamp;
 mod traits;
 mod types;
 
-#[allow(deprecated)]
-pub use crate::addresses::{Addr, CanonicalAddr, HumanAddr};
+pub use crate::addresses::{Addr, CanonicalAddr};
 pub use crate::binary::Binary;
 pub use crate::coins::{coin, coins, has_coins, Coin};
 pub use crate::deps::{Deps, DepsMut, OwnedDeps};